+++ src/env_node.mli
@@ -19,10 +19,16 @@ val external_ : t -> profile:string -> default:Env.t -> Env.t
 
 val ocaml_flags : t -> profile:string -> expander:Expander.t -> Ocaml_flags.t
 
-val c_flags : t -> profile:string -> expander:Expander.t
+val c_flags
+  : t
+  -> profile:string
+  -> expander:Expander.t
   -> (unit, string list) Build.t
 
-val cxx_flags : t -> profile:string -> expander:Expander.t
+val cxx_flags
+  : t
+  -> profile:string
+  -> expander:Expander.t
   -> (unit, string list) Build.t
 
 val local_binaries